# optimized _is_missing_literal