--- a/news/532.misc
+++ b/news/532.misc
@@ -1,1 +1,1 @@
-Optimized ListConfig iteration by 10.25x in a benchmark
+Optimized ListConfig iteration by 12x in a benchmark