Further speedup of ListConfig iteration.
```
                                        master     now     speedup
test__is_missing_literal                439.6 ns   96.0 ns    4.5x
test__list_iter[small_listconfig]       11.04 us   9.11 us    1.2x
```